--- app/_components/index.ts
@@ -1,6 +1,0 @@
-export { default as DisplayProducts } from './DisplayProducts';
-export { default as FilterChipsBox } from './FilterChipsBox';
-export { default as GlobalNavigationBar } from './GlobalNavigationBar';
-export { default as SideNavigationBar } from './SideNavigationBar';
-export { default as FloatingUpButton } from './FloatingUpButton';
-export { default as ProductDetail } from './ProductDetail';
