+++ app/components/Product/ProductContents.tsx
@@ -1,5 +1,5 @@
-import { GetProductListResponse } from '@/app/types';
-import { gruppo } from '@/app/utils/constants';
+import { GetProductListResponse } from '@/app/modules/types';
+import { gruppo } from '@/app/modules/constants';
 import { Box } from '@mui/material';
 import clsx from 'clsx';
