+++ app/components/common/Header.tsx
@@ -1,8 +1,8 @@
 'use client';
 
+import { gruppo } from '@/app/modules/constants';
 import { Box, Typography } from '@mui/material';
 import clsx from 'clsx';
-import { gruppo } from '../utils/constants';
 
 /**
  * 레이아웃 > 헤더 컴포넌트
