+++ app/components/common/SkeletonUI.tsx
@@ -5,35 +5,35 @@ import { Skeleton } from '@mui/material';
  */
 function SkeletonUI() {
   return (
-    <div className="h-96">
+    <div>
       <Skeleton
         variant="rectangular"
-        width={144}
-        height={216}
+        width="13rem"
+        height="15rem"
         className="m-auto my-3"
       />
       <Skeleton
         variant="rectangular"
-        width={250}
-        height={32}
+        width="16rem"
+        height="2rem"
         className="my-3"
       />
       <Skeleton
         variant="rectangular"
-        width={120}
-        height={24}
+        width="10rem"
+        height="2rem"
         className="my-2"
       />
       <Skeleton
         variant="rectangular"
-        width={150}
-        height={24}
+        width="14rem"
+        height="2rem"
         className="my-2"
       />
       <Skeleton
         variant="rectangular"
-        width={100}
-        height={24}
+        width="6rem"
+        height="2rem"
         className="my-2"
       />
     </div>
