--- app/ui/index.ts
@@ -1,9 +1,0 @@
-export { default as FilterChip } from './FilterChip';
-export { default as SNBFilterList } from './SNBFilterList';
-export { default as SkeletonUI } from './SkeletonUI';
-export { default as Header } from './Header';
-
-export { default as ProductColor } from './Product/ProductColor';
-export { default as ProductInfo } from './Product/ProductInfo';
-export { default as ProductContents } from './Product/ProductContents';
-export { default as ProductOverview } from './Product/ProductOverview';
